improve testing of delta decoding
