+++ encoding/delta/delta_test.go
@@ -4,10 +4,12 @@
 package delta_test
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/segmentio/parquet-go/encoding/delta"
 	"github.com/segmentio/parquet-go/encoding/fuzz"
+	"github.com/segmentio/parquet-go/encoding/test"
 )
 
 func FuzzDeltaBinaryPackedInt32(f *testing.F) {
@@ -24,4 +26,35 @@ func FuzzDeltaLengthByteArray(f *testing.F) {
 
 func FuzzDeltaByteArray(f *testing.F) {
 	fuzz.EncodeByteArray(f, new(delta.ByteArrayEncoding))
+}
+
+const (
+	encodeMinNumValues = 0
+	encodeMaxNumValues = 200
+)
+
+func TestEncodeInt32(t *testing.T) {
+	for bitWidth := uint(0); bitWidth <= 32; bitWidth++ {
+		t.Run(fmt.Sprintf("bitWidth=%d", bitWidth), func(t *testing.T) {
+			test.EncodeInt32(t,
+				new(delta.BinaryPackedEncoding),
+				encodeMinNumValues,
+				encodeMaxNumValues,
+				bitWidth,
+			)
+		})
+	}
+}
+
+func TestEncodeInt64(t *testing.T) {
+	for bitWidth := uint(0); bitWidth <= 64; bitWidth++ {
+		t.Run(fmt.Sprintf("bitWidth=%d", bitWidth), func(t *testing.T) {
+			test.EncodeInt64(t,
+				new(delta.BinaryPackedEncoding),
+				encodeMinNumValues,
+				encodeMaxNumValues,
+				bitWidth,
+			)
+		})
+	}
 }
